==> notebook/Stage01_DataCleaning.ipynb <==
import pandas as pd
import os

def create_df(file,dir_path):
    """This will read the file and return the dataframe
        provide the file which needs to be read
    """
    try:
        return pd.read_table(os.path.join(dir_path,file), delimiter=',')
    except:
        return "couldnt create datframe"

# all .data and .test need to be read 
dir_path = '../raw_data_directory'
allfilesl = [file for file in os.listdir(dir_path) if file.endswith('.data')| file.endswith('.test')]
df_allbp = create_df(allfilesl[0],dir_path)
df_allbptest = create_df(allfilesl[1],dir_path)
df_allhyper = create_df(allfilesl[2],dir_path)#done
df_allhypertest = create_df(allfilesl[3],dir_path)#done
df_allhypo = create_df(allfilesl[4],dir_path)#done
df_allhypotest = create_df(allfilesl[5],dir_path)#done
df_allrep = create_df(allfilesl[6],dir_path)
df_allreptest = create_df(allfilesl[7],dir_path)

df_dis = create_df(allfilesl[10],dir_path)
df_distest = create_df(allfilesl[11],dir_path)
df_hypothyroid = create_df(allfilesl[12],dir_path)#done
df_newthyroid = create_df(allfilesl[13],dir_path)
df_sickeuthyroid = create_df(allfilesl[14],dir_path)#done
df_sick = create_df(allfilesl[15],dir_path)
df_sicktest = create_df(allfilesl[16],dir_path)
df_thyroid0387 = create_df(allfilesl[17],dir_path)#done


columns = ['age','sex','on_thyroxine','query_on_thyroxine','on_antithyroid_medication','sick','pregnant','thyroid_surgery','I131_treatment','query_hypothyroid','query_hyperthyroid','lithium','goitre','tumor','hypopituitary','psych','TSH_measured','TSH','T3_measured','T3','TT4_measured','TT4','T4U_measured','T4U','FTI_measured','FTI','TBG_measured','TBG','referral_source','Target']
df_allhyper.columns = columns
df_allhyper.head()

df_allhypertest.columns = columns
df_allhypertest.head()

df_allhypo.columns = columns
df_allhypo.head()

df_allhypotest.columns = columns
df_allhypotest.head()

# Target column Cleaning

# removing the numerical value from the end of the Target column
def removenumber(column):
    for i in column:
        i = i.split('.')[0]
        return i

df_allhyper['Class']=df_allhyper[['Target']].apply(removenumber, axis=1)
df_allhypertest['Class']=df_allhypertest[['Target']].apply(removenumber, axis=1)
df_allhypo['Class']=df_allhypo[['Target']].apply(removenumber, axis=1)
df_allhypotest['Class']=df_allhypotest[['Target']].apply(removenumber, axis=1)

print("the number of unique values in Target coloumn of df_allhyper is ", df_allhyper['Class'].nunique(),"\nThey are" ,df_allhyper['Class'].value_counts().index)
print("the number of unique values in Target coloumn of df_allhypertest is ", df_allhypertest['Class'].nunique(),"\nThey are" ,df_allhypertest['Class'].value_counts().index)
print("the number of unique values in Target coloumn of df_allhypo is ", df_allhypo['Class'].nunique(),"\nThey are" ,df_allhypo['Class'].value_counts().index)
print("the number of unique values in Target coloumn of df_allhypotest is ", df_allhypotest['Class'].nunique(),"\nThey are" ,df_allhypotest['Class'].value_counts().index)

# Findings
- df_allhyper and df_allhypertest present classes---- ['negative', 'hyperthyroid', 'T3 toxic', 'goitre','secondary toxic']
- df_allhypo and df_allhypotest present classes------ ['negative', 'compensated hypothyroid', 'primary hypothyroid','secondary hypothyroid']

###  we will keep only [negative,hyperthyroid] in df_allhyper and df_allhypertest target column and [negative,hypothyroid] in df_allhypo and df_allhypotest target column

def cleanclasscolumn(df, given_target_values,target_true):
    df_negative = df[df.Class=='negative']
    df_cleaning = df[df.Class.isin(given_target_values)]
    df_cleaning.replace(given_target_values,target_true, inplace =True)
    df=pd.concat([df_negative,df_cleaning],ignore_index = True)
    return df

df_allhyper = cleanclasscolumn(df_allhyper, ['hyperthyroid', 'T3 toxic', 'goitre','secondary toxic'],'hyperthyroid' )
df_allhypertest = cleanclasscolumn(df_allhypertest, ['hyperthyroid', 'T3 toxic', 'goitre','secondary toxic'],'hyperthyroid')
df_allhypo = cleanclasscolumn(df_allhypo, ['compensated hypothyroid', 'primary hypothyroid','secondary hypothyroid'],'hypothyroid')
df_allhypotest = cleanclasscolumn( df_allhypotest, ['compensated hypothyroid', 'primary hypothyroid','secondary hypothyroid'],'hypothyroid')

# Post Cleaning of the Class Column

print("the number of unique values in Target coloumn of df_allhyper is ", df_allhyper['Class'].nunique(),"\nThey are" ,df_allhyper['Class'].value_counts().index)
print("the number of unique values in Target coloumn of df_allhypertest is ", df_allhypertest['Class'].nunique(),"\nThey are" ,df_allhypertest['Class'].value_counts().index)
print("the number of unique values in Target coloumn of df_allhypo is ", df_allhypo['Class'].nunique(),"\nThey are" ,df_allhypo['Class'].value_counts().index)
print("the number of unique values in Target coloumn of df_allhypotest is ", df_allhypotest['Class'].nunique(),"\nThey are" ,df_allhypotest['Class'].value_counts().index)

df_hh = pd.concat([df_allhyper,df_allhypertest,df_allhypo,df_allhypotest], ignore_index = True)
df_hh.head()

# shape of the column after concatenating the dataframes and before dropping duplicate rows
print(df_hh.shape)
# dropping the duplicate columns
df_hh.drop_duplicates(inplace =True)
# Dropping target column
df_hh.drop(columns = ['Target'], inplace =True)
# shape of the column after concatenating the dataframes and dropping duplicate rows
print(df_hh.shape)


# df_thyroid0387

df_thyroid0387.columns = columns
df_thyroid0387.head()

# Target Column Analysis
# cleaning Target Column
Class = []
for records in df_thyroid0387.Target:
    Class.append(str(records).split("[")[0])
# Adding the as Class Column into the dataframe
df_thyroid0387['Class'] = pd.Series(Class)
# Unique values in Class/Target column of df_thyroid0387
df_thyroid0387.Class.value_counts().index


# Set A,B,C,D to hyperthyroid
# Set E,F,G,H to hypothyroid
# Rest all classes will be set to negative
df_thyroid0387['Class'].replace(['A','B','C','D'],"hyperthyroid",inplace = True)
df_thyroid0387['Class'].replace(['E','F','G','H'],"hypothyroid",inplace = True)
for classes in df_thyroid0387['Class'].value_counts().index:
    if classes != "hyperthyroid" and classes !="hypothyroid":
        df_thyroid0387['Class'].replace(classes,"negative",inplace = True)
df_thyroid0387.Class.value_counts().index

df_thyroid0387.drop(columns = ["Target"], inplace =True)
df_thyroid0387.drop_duplicates(inplace =True)
print(df_thyroid0387.shape)
df_thyroid0387.head()

# df_hypothyroid

column_list = ['Unnamed','Age','Sex','on_thyroxine','query_on_thyroxine','on_antithyroid_medication','thyroid_surgery','query_hypothyroid','query_hyperthyroid','pregnant','sick','tumor','lithium','goitre','TSH_measured','TSH','T3_measured','T3','TT4_measured','TT4','T4U_measured','T4U','FTI_measured','FTI','TBG_measured','TBG']
print(len(column_list),df_hypothyroid.shape)
df_hypothyroid.columns =column_list
df_hypothyroid.head()

df_hypothyroid.Unnamed.value_counts().index

- df_hypothyroid.Unnamed is our target/Class column

# this dataset has only 26 columns while the previously observed dataset has 30 columns 
set_df_hh=   set(df_hh.columns) 
set_df_hypothyroid = set(df_hypothyroid.columns)
set_df_hh.difference(set_df_hypothyroid)

# we have 'Sex' and 'Age'column which we have to make 'sex' and 'age'
# df_hypothyroid.Unnamed is our target/Class column we have to rename it
df_hypothyroid = df_hypothyroid.rename(columns = {
    df_hypothyroid.columns[0]: "Class",
    df_hypothyroid.columns[1] :  "age",
    df_hypothyroid.columns[2] : "sex"
                                 })
df_hypothyroid

# df_sickeuthyroid

column_list = ['Unnamed','Age','Sex','on_thyroxine','query_on_thyroxine','on_antithyroid_medication','thyroid_surgery','query_hypothyroid','query_hyperthyroid','pregnant','sick','tumor','lithium','goitre','TSH_measured','TSH','T3_measured','T3','TT4_measured','TT4','T4U_measured','T4U','FTI_measured','FTI','TBG_measured','TBG']
print(len(column_list))
df_sickeuthyroid.columns =column_list
df_sickeuthyroid.head()

# this dataset has only 26 columns while the previously observed dataset has 30 columns 
set_df_hh=   set(df_hh.columns) 
set_df_sickeuthyroid = set(df_sickeuthyroid.columns)
set_df_hh.difference(set_df_sickeuthyroid)

# we have 'Sex' and 'Age'column which we have to make 'sex' and 'age'
# df_hypothyroid.Unnamed is our target/Class column we have to rename it
df_sickeuthyroid = df_sickeuthyroid.rename(columns = {
    df_sickeuthyroid.columns[0]: "Class",
    df_sickeuthyroid.columns[1] :  "age",
    df_sickeuthyroid.columns[2] : "sex"
                                 })
df_sickeuthyroid

# df_anntrain

# df_anntest

columns_ann = ["age","sex" ,"on_thyroxine", "query_on_thyroxine" ,  "on_antithyroid_medication" , "sick", "pregnant","thyroid_surgery" ,"I131_treatment"  ,"query_hypothyroid" ,"query_hyperthyroid" ,"lithium","goitre" , "tumor" ,"hypopituitary","psych","TSH"  ,"T3" ,"TT4", "T4U" ,"FTI", "Class" ]                                             

df_anntest = pd.read_csv(os.path.join(dir_path,"ann-test.data"))
df_anntrain = pd.read_csv(os.path.join(dir_path,"ann-train.data"))
df_anntest

#all the data is clubbed together in a single column seperated by " "


df_anntest = df_anntest.iloc[:,0].str.split(' ', expand=True)
df_anntrain = df_anntrain.iloc[:,0].str.split(' ', expand=True)

df_anntest.head()

df_anntrain.head()

# both of these have last wo columns empty we can delete that
df_anntrain = df_anntrain.drop(columns = [22,23])
df_anntest  = df_anntest.drop(columns = [22,23])

df_anntest.columns = columns_ann
df_anntrain.columns = columns_ann

for col in df_anntrain.columns:
    df_anntrain[col] = pd.to_numeric(df_anntrain[col])
    df_anntest[col] =  pd.to_numeric(df_anntest[col])

target1 = pd.Series(df_anntest[df_anntest.columns[-1]].values)
display(target1.value_counts())
target2 = pd.Series(df_anntrain[df_anntrain.columns[-1]].values)
display(target2.value_counts())

- 3 ==  'negative' class
- 2 == 'hypothyroid' class
- 1 == 'hyperthyroid' class

At this point we are not sure whether 0 reresents male or 1
We need to check other dataset for sex column distribution

print("Sex thyroid0387 1=F,0=M:")
sex_series1 = pd.Series(df_thyroid0387[df_thyroid0387.columns[1]].values)
display(sex_series1.value_counts())
print("Sick-euthyroid:")
sex_series2 = pd.Series(df_sickeuthyroid[df_sickeuthyroid.columns[2]].values)
display(sex_series2.value_counts())

####  Female are more than male
#### Another important things to do is to multply for 100 all the continuos and numerical attributes and to add the 'measured' attributes. Looking at age column it is represented in decimals

df_anntest.sex.value_counts()

#### 0 represents F and 1 M in ann dataset

for column in df_anntest.columns:
    list_Values=set(df_anntest[column])
    print(column,": ",list_Values)

ann = pd.concat([df_anntrain,df_anntest], ignore_index = True)
ann['sex'] = ann['sex'].map({0:'F',1:'M'})
ann['Class'] = ann['Class'].map({3:'negative',2:'hypothyroid',1:'hyperthyroid'})

continuos_attributes = ['age','TSH','T3','TT4','T4U','FTI']
for attribute in continuos_attributes:
    ann[attribute] = ann[attribute] * 100

def fillNewAttributes(row,attribute):
    if row[attribute] > 0:
        return 'y'
    else:
        return 'n'

ann['TSH_measured'] = ann.apply(lambda row: fillNewAttributes(row,'TSH'), axis=1)
ann['T3_measured'] = ann.apply(lambda row: fillNewAttributes(row,'T3'), axis=1)
ann['TT4_measured'] = ann.apply(lambda row: fillNewAttributes(row,'TT4'), axis=1)
ann['T4U_measured'] = ann.apply(lambda row: fillNewAttributes(row,'T4U'), axis=1)
ann['FTI_measured'] = ann.apply(lambda row: fillNewAttributes(row,'FTI'), axis=1)
display(ann.dtypes)

# Findings
- df_allbp,df_allhyper,df_allhypo,df_allrep,df_dis,df_sick dataframes seem to follow same structure- we will verify this and merge all records
- df_anntrain,df_hypothyroid,df_sickeuthyroid do not share same structure
- 
- df_hypothyroid present classes 'hypothyroid' and 'negative'
- df_sickeuthyroid present classes 'sick-euthyroid' and 'negative'
- df_thyroid0387 present classes hyperthyroid conditions (A, B, C, D), hypothyroid conditions (E, F, G, H), binding protein (I, J), general health (K), replacement therapy (L, M, N), discordant results (R)
- df_anntest and df_anntrain present classes normal (not hypothyroid), hyperfunction and subnormal functioning

data = pd.concat([df_hh,df_thyroid0387,df_hypothyroid,df_sickeuthyroid,ann], ignore_index = True)
display(data.shape)
display(data.dtypes)

dir_path_post = '../cleaned_processed_data'
data.to_csv(os.path.join(dir_path_post,'stage01_cleaned_data.csv'), index = False)
