==> notebook/Stage02_DataPreprocessing.ipynb <==
import pandas as pd
import os
import numpy as np

dir_path_post = '../cleaned_processed_data'
df = pd.read_csv(os.path.join(dir_path_post,'stage01_cleaned_data.csv'))
df.head()

# unique values of the data set

for col in df.columns:
    print(col)
    print("*****************************")
    print(df[col].value_counts())

## there are "?" entries which denotes missingvalues or nan values replacing them
df=df.replace({"?":np.NAN})
df.isna().sum()

## Dropping the rows with more than 10 nan values and columns with more than 7000 nans 
def dropna_thresh(dataframe):
    dataframe.dropna(thresh=20, inplace =True)
    column_drop = []
    for col in dataframe.columns:
        if dataframe[col].isna().sum() > 7000:
            column_drop.append(col)
    dataframe = dataframe.drop(columns = column_drop)
    return dataframe
df = dropna_thresh(df)
df.isna().sum()

### Encoding into numerical values 
def encode(dataframe):
    """ 
    1. sex column has 'F' and 'M' entries
    2. on_thyroxine,query_on_thyroxine,on_antithyroid_medication,sick,pregnant
thyroid_surgery,I131_treatment,query_hypothyroid,lithium,goitre,tumor,hypopituitary
psych has 'f' and 't entries'
3. TSH_measured,T3_measured,TT4_measured,T4U_measured,FTI_measured has 'f','n','t','y' entries
    """
    dataframe['sex'] = dataframe['sex'].replace({'F':1, 'M':0})
    dataframe = dataframe.replace({"t":1,"f":0})
    dataframe = dataframe.replace({"y":1, "n":0})
    dataframe['Class'] = dataframe['Class'].replace({"negative":0,"hypothyroid":1,"hyperthyroid":2,"sick-euthyroid":3})
    for col in dataframe.columns:
        try:
            dataframe[col] = dataframe[col].apply(pd.to_numeric , errors = 'coerce')
        except:
            pass
    return dataframe
    

df = encode(df)
df.dtypes


df.head()

display(df.shape)
df['Class'].value_counts()

#### it seems like data is highly imbalanced . if its true then we cannot take this data forward without imputing the nan 


data1 = df.interpolate(method = 'spline', order = 3)
display(data1.isna().sum())

data1.dropna(inplace =True)

data1.to_csv(os.path.join(dir_path_post,'stage02_preprocessed.csv'), index = False)
