==> notebook/Stage03_ModelTraining.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
import os
from catboost import CatBoostClassifier
from  xgboost import XGBClassifier
from sklearn.tree import DecisionTreeClassifier 
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import ExtraTreesClassifier
from sklearn import metrics
from sklearn.metrics import confusion_matrix
from sklearn.preprocessing import LabelEncoder
from sklearn.preprocessing import MinMaxScaler
from imblearn.over_sampling import SMOTE
from sklearn.utils import resample

dir_path_post = '../cleaned_processed_data'
df = pd.read_csv(os.path.join(dir_path_post,'stage02_preprocessed.csv'))
df.head(5)

- features which are highly correlated with Class label... 

def get_correlation(df):
    corr_values = abs(df.corr(method='spearman')['Class'])
    corr_values = corr_values.drop('Class')
    corr_values = corr_values[corr_values > 0.03]
    display(corr_values)
    return corr_values


corr_values = get_correlation(df)

### Train Test Split

def traintest_data(df, features, label):
    X = df[features]
    Y = df[label]
    X_train, X_test, y_train, y_test = train_test_split(X,Y, test_size=0.30, random_state=42) 
    return X_train, X_test, y_train, y_test


X_train, X_test, y_train, y_test = traintest_data(df, corr_values.index, 'Class')

# Classifiers we are going to use
classifiers = {
    "XGBClassifier" : XGBClassifier(learning_rate=0.01),
    "Nearest Neighbors" : KNeighborsClassifier(4),
    "Decision Tree" : DecisionTreeClassifier(class_weight = 'balanced'),
    "Random Forest": RandomForestClassifier(class_weight = 'balanced',random_state = 1),
    "ExtraTrees": ExtraTreesClassifier(class_weight = 'balanced',random_state = 1),
    "CatBoostClassifier" : CatBoostClassifier(max_depth=4,verbose=0),
}

def train_getscore(classifiers,X_train, X_test, y_train, y_test ):
    score_board = pd.DataFrame(columns = ["Classifier", 
                                "Accuracy", 
                                "Precision", 
                                "Recall", 
                                "FScore"])
    for name, clf in classifiers.items():
        clf.fit(X_train,y_train)
        y_pred = clf.predict(X_test)
        
        precision_score,recall_score,f_score,support = metrics.precision_recall_fscore_support(y_test,y_pred,average='macro')
        score_board=score_board.append({ "Classifier": name,
                           "Accuracy": round(metrics.accuracy_score(y_test, y_pred), 4),
                           "Precision": round(precision_score, 4), 
                           "Recall":round(recall_score, 4), 
                           "FScore":round(f_score, 4)
                           }, ignore_index=True)
        print("Confusion matrix for: ", name)
        display(confusion_matrix(y_test, y_pred))
    score_board.set_index("FScore", inplace=True)
    score_board.sort_values(by="FScore", ascending=False, inplace=True)   
    return score_board
        
display(train_getscore(classifiers, X_train, X_test, y_train, y_test))        

# it seems like data is highly imbalanced . if its true then we cannot only rely on 
display(df.shape)
df['Class'].value_counts()

# https://www.javatpoint.com/discretization-in-data-mining

def fdiscretizer(attribute,dataframe):
    enc = LabelEncoder()
    dataframe[attribute] = pd.qcut(dataframe[attribute], 20, duplicates='drop')
    dataframe[attribute] = enc.fit_transform(dataframe[attribute])
    dataframe = dataframe.convert_dtypes(convert_integer=True)

data2 = df.copy()
fdiscretizer('age',data2)
fdiscretizer('TSH',data2)
fdiscretizer('T3',data2)
fdiscretizer('TT4',data2)
fdiscretizer('T4U',data2)
fdiscretizer('FTI',data2)

corr_values = get_correlation(data2)
X_train2, X_test2, y_train2, y_test2 = traintest_data(data2, corr_values.index,'Class')

display(train_getscore(classifiers,X_train2, X_test2, y_train2, y_test2))

### F1 score has dropped 
###  Might be discretization is not a valid option
###  will check Standardisation
### T3 feature is now  seems to be highly correlted with Class

scaler = MinMaxScaler()
data3=pd.DataFrame(scaler.fit_transform(df.drop(columns =['Class'])),columns=data2.columns[:-1])
data3['Class'] = df['Class']
# data3 = ((df-df.min())/(df.max()-df.min()))*20

corr_values = get_correlation(data3)
X_train3, X_test3, y_train3, y_test3 = traintest_data(data3, corr_values.index,'Class')

display(train_getscore(classifiers,X_train3, X_test3, y_train3, y_test3))

### f1 score has risen from last 
### still the data is imbalanced need to balance the data

smote = SMOTE('not majority',random_state = 1)
X_train_sm, y_train_sm = smote.fit_resample(X_train3,y_train3)
X_test_sm, y_test_sm = smote.fit_resample(X_test3,y_test3)
display(X_train3.shape)
display(X_train_sm.shape)
display(train_getscore(classifiers,X_train_sm, X_test_sm, y_train_sm, y_test_sm))


## Under sampling method
df_negative = data3[data3.Class==0]
df_hypothyroid = data3[data3.Class==1]
df_hyperthyroid = data3[data3.Class==2]
df_sickeuthyroid = data3[data3.Class==3]

df_negative_downsampled = resample(df_negative,replace=False,n_samples=450,random_state=1)
df_hypothyroid_downsampled = resample(df_hypothyroid,replace=False,n_samples=450,random_state=1)
df_hyperthyroid_downsampled = resample(df_hyperthyroid,replace=False,n_samples=450,random_state=1)
#df_sickeuthyroid_downsampled = resample(df_sickeuthyroid,replace=False,n_samples=450,random_state=1)



df_downsampled = pd.concat([df_negative_downsampled,df_hypothyroid_downsampled,df_hyperthyroid,df_sickeuthyroid])
df_downsampled.Class.value_counts()

X_train4, X_test4, y_train4, y_test4 = traintest_data(df_downsampled,corr_values.index,'Class')
display(train_getscore(classifiers,X_train4, X_test4, y_train4, y_test4))

# final for generating pickle file

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
import os
from sklearn.ensemble import RandomForestClassifier
import pickle
from sklearn import metrics
from sklearn.metrics import confusion_matrix
from sklearn.preprocessing import LabelEncoder
from sklearn.preprocessing import MinMaxScaler
from imblearn.over_sampling import SMOTE
from sklearn.utils import resample
import gzip, pickle, pickletools

dir_path_post = '../cleaned_processed_data'
df = pd.read_csv(os.path.join(dir_path_post,'stage02_preprocessed.csv'))
df.head(5)



def get_correlation(df):
    corr_values = abs(df.corr(method='spearman')['Class'])
    corr_values = corr_values.drop('Class')
    corr_values = corr_values[corr_values > 0.03]
    display(corr_values)
    return corr_values

def traintest_data(df, features, label):
    X = df[features]
    Y = df[label]
    X_train, X_test, y_train, y_test = train_test_split(X,Y, test_size=0.30, random_state=42) 
    return X_train, X_test, y_train, y_test


def scaled_df(df):
    scaler = MinMaxScaler()
    
    data=pd.DataFrame(scaler.fit_transform(df.drop(columns =['Class'])),columns=df.columns[:-1])
    pickle.dump(scaler, open('../model_scaler_picklefile/thyroid_scaler.pickle','wb'))
    data['Class'] = df['Class']
    return data

def smote_resampled(X_train, X_test, y_train, y_test):
    smote = SMOTE('not majority',random_state = 1)
    X_train_sm, y_train_sm = smote.fit_resample(X_train,y_train)
    X_test_sm, y_test_sm = smote.fit_resample(X_test,y_test)
    return X_train_sm,X_test_sm,y_train_sm,y_test_sm


def under_sampled(df):
    df_negative = df[df.Class==0]
    df_hypothyroid = df[df.Class==1]
    df_hyperthyroid = df[df.Class==2]
    df_sickeuthyroid = df[df.Class==3]

    df_negative_downsampled = resample(df_negative,replace=False,n_samples=450,random_state=1)
    df_hypothyroid_downsampled = resample(df_hypothyroid,replace=False,n_samples=450,random_state=1)
    df_hyperthyroid_downsampled = resample(df_hyperthyroid,replace=False,n_samples=450,random_state=1)
    
    df_downsampled = pd.concat([df_negative_downsampled,df_hypothyroid_downsampled,df_hyperthyroid,df_sickeuthyroid])
    return df_downsampled

data = df.copy()

#### List of selected features
selected_features = ['age', 'sex', 'on_thyroxine', 'query_hypothyroid', 'query_hyperthyroid',
       'psych', 'TSH_measured', 'TSH', 'T3_measured', 'T3', 'TT4_measured',
       'TT4', 'T4U_measured', 'FTI_measured', 'FTI']
all_data = ['age', 'sex', 'on_thyroxine', 'query_hypothyroid', 'query_hyperthyroid',
       'psych', 'TSH_measured', 'TSH', 'T3_measured', 'T3', 'TT4_measured',
       'TT4', 'T4U_measured', 'FTI_measured', 'FTI','Class']
data = scaled_df(df[all_data])
df_downsampled = under_sampled(data[all_data])
X_train_final, X_test_final, y_train_final, y_test_final = traintest_data(df_downsampled,selected_features,'Class')
rfc = RandomForestClassifier(class_weight = 'balanced',random_state = 10)
rfc.fit(X_train_final,y_train_final)
y_pred = rfc.predict(X_test_final)

precision_score,recall_score,f_score,support = metrics.precision_recall_fscore_support(y_test_final,y_pred,average='macro')
precision_score,recall_score,f_score

# discretization_columns = ['age','TSH','T3','TT4','T4U','FTI']
data = df.copy()
# for col in discretization_columns:
#     data[col] = fdiscretizer(col,data)
# scaler = MinMaxScaler()
# data=pd.DataFrame(scaler.fit_transform(df.drop(columns =['Class'])),columns=data2.columns[:-1])
# data['Class'] = df['Class']

data = scaled_df(df)

corr_values = get_correlation(data)
X_train, X_test, y_train, y_test = traintest_data(data, corr_values.index,'Class')

X_train_sm,X_test_sm,y_train_sm,y_test_sm = smote_resampled(X_train, X_test, y_train, y_test)
df_downsampled = under_sampled(data)
X_train_final, X_test_final, y_train_final, y_test_final = traintest_data(df_downsampled,corr_values.index,'Class')



# Model
rfc = RandomForestClassifier(class_weight = 'balanced',random_state = 10)
rfc.fit(X_train_final,y_train_final)
y_pred = rfc.predict(X_test_final)

precision_score,recall_score,f_score,support = metrics.precision_recall_fscore_support(y_test_final,y_pred,average='macro')

# model pickle

filepath = os.path.join("../model_scaler_picklefile","thyroid_model.pkl")
with gzip.open(filepath, "wb") as f:
    pickled = pickle.dumps(rfc)
    optimized_pickle = pickletools.optimize(pickled)
    f.write(optimized_pickle)

#### List of selected features
selected_features = ['age', 'sex', 'on_thyroxine', 'query_hypothyroid', 'query_hyperthyroid',
       'psych', 'TSH_measured', 'TSH', 'T3_measured', 'T3', 'TT4_measured',
       'TT4', 'T4U_measured', 'FTI_measured', 'FTI']

df[selected_features].iloc[1,:].values
